feat: extract bet data module
