+++ index.html
@@ -123,6 +123,6 @@
   <div id="include-modal"></div>
 
   <script src="js/loadShared.js"></script>
-  <script src="js/app.js"></script>
+  <script type="module" src="js/app.js"></script>
 </body>
 </html>
